team, teaminfo 추가 테스트 완료.

+++ src/main/java/jpabasic/reserve/domain/TestService.java
@@ -1,5 +1,8 @@
 package jpabasic.reserve.domain;
 
+import java.util.Map;
+import java.util.Objects;
+
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityManagerFactory;
 import jakarta.persistence.EntityTransaction;
@@ -11,6 +14,7 @@ public abstract class TestService {
 	public static final String FIND    = "find"; 
 	public static final String CHANGE  = "change"; 
 	public static final String REMOVE  = "remove"; 
+	public static final String INIT    = "init"; 
 	
 	public void persist(EntityManager manager, EntityTransaction transaction) {}
 	
@@ -20,6 +24,8 @@ public abstract class TestService {
 	
 	public void remove(EntityManager manager, EntityTransaction transaction) {}
 	
+	public void init(EntityManager manager, EntityTransaction transaction) {}
+	
 	
 	public static void test(String name, TestService service) {
 		EntityManagerFactory factory = Persistence.createEntityManagerFactory("jpabegin");;
@@ -27,20 +33,35 @@ public abstract class TestService {
 		EntityTransaction transaction = manager.getTransaction();
 		
 		try {
-			System.out.println("############### test_"+ name +" start");
-			transaction.begin();
-			System.out.println("\n\n");
+			Map<String, Object> properties = factory.getProperties();
+			String ddl = Objects.toString(properties.get("hibernate.hbm2ddl.auto"), "");
 			
-			switch(name) {
-				case PERSIST: service.persist(manager, transaction); break;
-				case FIND   : service.find(manager, transaction); break;
-				case CHANGE : service.change(manager, transaction); break;
-				case REMOVE : service.remove(manager, transaction); break;
+			if("create".equals(ddl)) {
+				System.out.println("\nONLY INIT TEST");
+				System.out.println(" - please change 'hibernate.hbm2ddl.auto' to 'update'\n");
+				
+				service.init(manager, transaction);
+				
+			} else {
+				
+				System.out.println("############### test_"+ name +" start");
+				transaction.begin();
+				System.out.println("\n\n");
+				
+				switch(name) {
+					case PERSIST: service.persist(manager, transaction); break;
+					case FIND   : service.find(manager, transaction); break;
+					case CHANGE : service.change(manager, transaction); break;
+					case REMOVE : service.remove(manager, transaction); break;
+					case INIT   : 
+						 default: service.init(manager, transaction); break;
+				}
+				
+				System.out.println("\n\n");
+				transaction.commit();
+				System.out.println("############### test_"+ name +" end");
 			}
 			
-			System.out.println("\n\n");
-			transaction.commit();
-			System.out.println("############### test_"+ name +" end");
         } catch (Exception ex) {
             ex.printStackTrace();
